Moves utils functions to proper file + bump version
