+++ VERSION
@@ -1,1 +1,1 @@
-0.7.0
+1.0.0
